refactor: TextView 및 ImageView Util 관리
- 분산되어 있는 TextView 및 ImageView의 관련 util을 하나로 통합해 관리한다.
