+++ app/src/main/java/com/woowa/banchan/ui/common/viewutils/ImageViewUtil.kt
@@ -1,0 +1,14 @@
+package com.woowa.banchan.ui.common.viewutils
+
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+
+@BindingAdapter("image")
+fun ImageView.setImage(url: String?) {
+    url?.let {
+        Glide.with(context)
+            .load(url)
+            .into(this)
+    }
+}
