+++ app/src/main/java/com/woowa/banchan/ui/common/viewutils/TextViewUtil.kt
@@ -1,6 +1,7 @@
 package com.woowa.banchan.ui.common.viewutils
 
 import android.graphics.Paint
+import android.view.View
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import com.woowa.banchan.ui.cart.cart.CartFragment
@@ -38,4 +39,26 @@ fun TextView.setCartFreeShipping(price: Int) {
 @BindingAdapter("viewedDate")
 fun TextView.setViewedTime(date: Date) {
     text = DateUtil.getUpdateDate(date)
+}
+
+@BindingAdapter("nPricePercent", "sPricePercent")
+fun TextView.setPercent(nPrice: String?, sPrice: String) {
+    if (nPrice == null) {
+        text = ""
+        visibility = View.GONE
+    } else {
+        val originPrice = nPrice.replace("원", "").replace(",", "").toLong()
+        val salePrice = sPrice.replace("원", "").replace(",", "").toLong()
+        val percent = (((originPrice - salePrice) * 100) / originPrice)
+        text = "$percent%"
+    }
+}
+
+@BindingAdapter("nPrice", "sPrice")
+fun TextView.setOriginPrice(nPrice: String?, sPrice: String) {
+    paintFlags = Paint.STRIKE_THRU_TEXT_FLAG
+    if (nPrice == null)
+        text = ""
+    else
+        text = nPrice
 }
